Merge Ilan's main working branch up to now into main

+++ src/dataloader.py
@@ -238,7 +238,9 @@ class TiggeMRMSDataset(Dataset):
             X = self._make_sr_X(y)
         if self.cat_bins is not None and not no_cat:
             y = self._categorize(y)
-        return X, y   # [vars, patch, patch]
+            
+#         return X, y
+        return X.astype(np.float32), y.astype(np.float32)   # [vars, patch, patch]
     
     def _add_const(self, X, lat_slice, lon_slice):
         """Add constants to X"""
